requirements.txt
django==3.2
mysqlclient==2.0.3
numpy==1.19.5
pillow
boto3
requests
asgiref==3.3.3
gunicorn==20.0.4
pytz==2019.3
setproctitle==1.1.10
